Testing permalinks
I'm thinking of having two URLs for the same post:

The file structure will be strictly date based, and so a path of

    2018/01/05/index.md

will result in a url of 'romellem.github.io/the-riddler/2018/01/05/'. Keying off of the relevant
Riddler post at fivethirtyeight.com/features/where-will-the-seven-dwarfs-sleep-tonight/,
I can also use the slug of the post as my permalink, meaning then that
'romellem.github.io/the-riddler/where-will-the-seven-dwarfs-sleep-tonight/'
will also link to the same page.